+++ packages/web-components/src/components/cbp-segmented-button-group/cbp-segmented-button-group.tsx
@@ -43,14 +43,16 @@ export class SegmentedButtonGroup {
 
   @Listen('buttonClick')
   handleButtonClick({ detail: { nativeElement: element, host, value } }) {
-    // Toggle "pressed" prop on the button
-    host.pressed=!host.pressed;
+    // Toggle "pressed" prop on the button only when multiple selections are allowed; 
+    // otherwise, it acts like a radio list and a click selects it but doesn't deselect it.
+    if (this.multiple) host.pressed = (host.pressed == "true") ? "false" : "true";
+    else host.pressed = "true";
 
     // if a button was toggled to "pressed," toggle the rest unpressed for groups that only allow a single buttons pressed.
     if(!this.multiple && host.pressed) {
       this.buttongroup.forEach(el => {
         if(el != host){
-          el.pressed=false;
+          el.pressed="false";
         }
       });
     }
@@ -74,6 +76,16 @@ export class SegmentedButtonGroup {
       ...this.sx,
     });
   }
+
+  componentDidLoad() {
+    // Set the pressed state of each button to "false" if it's not set to "true" already.
+    this.buttongroup.forEach( cbpButton => {
+      if(cbpButton.pressed !== "true"){
+        cbpButton.pressed="false";
+      }
+    });
+  }
+
 
   render() {
     return (